alfred: Don't force cleanup of transaction on TXEND
A committed transaction has always to be cleaned up/removed when the TXEND
was received. But TXEND packets which are out-of-order may still be waiting
for PUSH_DATA packets which will be receive later. Thus TXEND packets which
don't trigger a commit of the PUSH_DATA should not automatically drop the
transaction.

A transaction which doesn't get all PUSH_DATA packets will be timeout
automatically and no extra handling is necessary.

+++ alfred.h
@@ -71,7 +71,7 @@ struct transaction_head {
 	struct ether_addr server_addr;
 	uint16_t id;
 	uint8_t requested_type;
-	int finished;
+	uint16_t finished;
 	int num_packet;
 	int client_socket;
 	struct timespec last_rx_time;
